Rename Cookie to Storage.
